atomisp2: css2401 and 2401_legacy-irci_master_20150113_1500
Merge CSS2.1 for css2401 and 2401_legacy
for irci_master_20150113_1500 firmware tag

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-16987
Reviewed-on: https://icggerrit.ir.intel.com/36205
Approver: Wei Tao <wei.tao@intel.com>

+++ drivers/external_drivers/camera/drivers/media/pci/atomisp2/css2401a0_v21/runtime/isp_param/interface/ia_css_isp_param_types.h
@@ -1,7 +1,7 @@
 /*
  * Support for Intel Camera Imaging ISP subsystem.
  *
- * Copyright (c) 2010 - 2014 Intel Corporation. All Rights Reserved.
+ * Copyright (c) 2010 - 2015 Intel Corporation. All Rights Reserved.
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License version
@@ -46,8 +46,8 @@ enum ia_css_param_class {
 
 /** ISP parameter descriptor */
 struct ia_css_isp_parameter {
-	uint16_t offset; /* Offset in isp_<mem>)parameters, etc. */
-	uint16_t size;   /* Disabled if 0 */
+	uint32_t offset; /* Offset in isp_<mem>)parameters, etc. */
+	uint32_t size;   /* Disabled if 0 */
 };
